WIP: Successful Google Auth, broke pretty much everything else in the process
