+++ .env.sample
@@ -14,9 +14,12 @@ DEPLOYMENT=self
 ENABLE_UPDATES=true
 DEBUG=sql,cache,presenters,events
 
-# Third party credentials (required)
+# Slack signin credentials (at least one is required)
 SLACK_KEY=71315967491.XXXXXXXXXX
 SLACK_SECRET=d2dc414f9953226bad0a356cXXXXYYYY
+
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
 
 # Third party credentials (optional)
 SLACK_VERIFICATION_TOKEN=PLxk6OlXXXXXVj3YYYY
