+++ package.json
@@ -99,6 +99,7 @@
     "file-loader": "^1.1.6",
     "flow-typed": "^2.4.0",
     "fs-extra": "^4.0.2",
+    "google-auth-library": "^1.5.0",
     "history": "3.0.0",
     "html-webpack-plugin": "2.17.0",
     "http-errors": "1.4.0",
@@ -168,11 +169,11 @@
     "styled-components-breakpoint": "^1.0.1",
     "styled-components-grid": "^1.0.0-preview.15",
     "styled-normalize": "^2.2.1",
+    "uglifyjs-webpack-plugin": "1.2.5",
     "url-loader": "^0.6.2",
     "uuid": "2.0.2",
     "validator": "5.2.0",
     "webpack": "3.10.0",
-    "uglifyjs-webpack-plugin": "1.2.5",
     "webpack-manifest-plugin": "^1.3.2"
   },
   "devDependencies": {
