+++ server/api/index.js
@@ -2,8 +2,6 @@
 import bodyParser from 'koa-bodyparser';
 import Koa from 'koa';
 import Router from 'koa-router';
-import Sequelize from 'sequelize';
-import _ from 'lodash';
 
 import auth from './auth';
 import user from './user';
@@ -16,58 +14,24 @@ import shares from './shares';
 import team from './team';
 import integrations from './integrations';
 
-import validation from './middlewares/validation';
-import methodOverride from '../middlewares/methodOverride';
-import cache from '../middlewares/cache';
+import errorHandling from './middlewares/errorHandling';
+import validation from '../middlewares/validation';
+import methodOverride from './middlewares/methodOverride';
+import cache from './middlewares/cache';
 import apiWrapper from './middlewares/apiWrapper';
 
 const api = new Koa();
 const router = new Router();
 
-// API error handler
-api.use(async (ctx, next) => {
-  try {
-    await next();
-  } catch (err) {
-    ctx.status = err.status || 500;
-    let message = err.message || err.name;
-    let error;
-
-    if (err instanceof Sequelize.ValidationError) {
-      // super basic form error handling
-      ctx.status = 400;
-      if (err.errors && err.errors[0]) {
-        message = `${err.errors[0].message} (${err.errors[0].path})`;
-      }
-    }
-
-    if (message.match('Authorization error')) {
-      ctx.status = 403;
-      error = 'authorization_error';
-    }
-
-    if (ctx.status === 500) {
-      message = 'Internal Server Error';
-      error = 'internal_server_error';
-      ctx.app.emit('error', err, ctx);
-    }
-
-    ctx.body = {
-      ok: false,
-      error: _.snakeCase(err.id || error),
-      status: err.status,
-      message,
-      data: err.errorData ? err.errorData : undefined,
-    };
-  }
-});
-
+// middlewares
+api.use(errorHandling());
 api.use(bodyParser());
 api.use(methodOverride());
 api.use(cache());
 api.use(validation());
 api.use(apiWrapper());
 
+// routes
 router.use('/', auth.routes());
 router.use('/', user.routes());
 router.use('/', collections.routes());
